Fix updates to ComponentPhraseModel.
Handle the case when descriptive phrases belonging to the
same parent item refer to the same persistent object. For
now, assume that the same object should only appear once
per parent. In the future, other phrase information should
be used to discriminate phrases.

This fixes the issue where the associated-entity list would
be cleared immediately after an operation ran -- not because
the operation removed existing objects, but because new phrases
were constructed that did not match the old phrase (causing a
removal and insertion of the same object, which reset the map
used to hold selected objects).