Query an entity's color status before fetching its color
In C++ if a key is not in the map, calling [] operator on it will create
a dummy entry. This will prevent other code to query the correct status
of the entity's color. By adding the right check code, an empty color property
will no longer be to the resource.