BUG:Fix Group View's Current Tab being reset
When changing either the category filter or advance level setting, nested group view would reset their current tab to the first one.
This can be very annoying.  This change now stores the view name associated with the current tab and stores it in the UI Manager in a map.
The key is the name of the group view and the value is the name of the currently selected child view.  When the group view is reconstructed,
this information is used to determine which tabs are current.