BUG: Items not updated if switching Attributes in Attribute View
If an QlineEdit box was being editting and the user then switched attributes from the top list, the corresponding Attribute Item would fail to update.
This was caused by the QTable's signals being processed before the QlineEdit's EdittingFinished signal - the Table's signals are now Queued so the
lineEdit signals can be directly processed first.