Ignore "decorated name length exceeded, name was truncated" Warning
Use of boost::mpl can result in some template instantations with very
long names. Windows truncates these names into something more
manageable with a warning to this effect. The internet seems to think
that this is an innocuous warning (unless it isn't; there can arise
linking issues, but we haven't seen these). The most common fix is to
simply disable these warnings in "I know what I'm doing" parts of the
code.