Fix method used to move phrases.
Note that std::vector<...>::insert() does not allow the source and
destination iterators to point to the same vector even when the
ranges do not overlap. Implementation-wise, this only exhibits
problems on platforms that use libstdc++ (i.e., Linux), where the
insertion may add null shared-pointer values instead of copies as
expected.