smtk::model::Entity: reject empty groups when querying components
When a query string does not explicitly request a group but does
request an entity type, empty groups are returned. For example, the
query string 'edge' will return all empty groups in addition to
edges. This MR changes the filter logic to exclude empty groups unless
groups are explicitly requested.