Use smtk::model::Group API to add entities
The method smtk::model::Resource::addToGroup() does not do the same
thing as smtk::model::Group::addEntity(), and the former results in a
group not having access to its entities. This MR simply changes the
test 'modelAttributes.py' to use the Group API, since this bug was
causing test failures once another bug was addressed.