ENH: Changing removeValue methods for items
Originally removeValue would simply return false if the item was not extensible. The new
behavior is if the index of the value to be removed is < the number of required values then
the method acts like unset - else it will remove the storage for the value as well (hence
number of values will be decrease).