smtk::mesh::moab::BufferedCellAllocator: enable multiple allocations
Originally, the smtk::mesh::moab implementation for the
BufferedCellAllocator would allow for a single allocation of points
and then iterative allocations of cells that used these points. This
was problematic when the allocator was used multiple times (which it
always is for a given resource, since it is cached by the moab
interface). This MR clears the state of the allocator each time it is
requested through the Interface API.