BUG: Highliting in Association Widget causes a crash
If a QLineEdit widget (generated from a qtValueItem) is being editted
and the user does not press Tab or Return to generate a editting finished signal but instead
proceeds to select objects from the Association Widget lists - you can generate a crash
if an item is selected and a previous item is trying to have its background set back to normal.

This appears to be a QT bug (currently using 5.12.3 on Mac) - this change
turns off the list highlighting mechanism.