BUG: Valid Attributes don't display properly in Attribute View
The issue was that it was assumed a QTabelWidget's selectItems() method would return
the all of the item pointers in a selected row even if they were nullptr (as in the case
of a valid attrubute - its status column would be nullptr).  That is not the case.
Instead you need to look at the returned items to find the one in column you are looking for.