BUG: Fixing Crash due to Project Close
Added new method qtItem::markForDeletion - this method should replace deleteLater calls.

The problem was Observers definined in qtItem derived classes such as qtReferenceItemComboBox were being called even when the qtItem was in the process of being deleted.  This new method (and the ability to override it) provides an
oppertunity for the qtItem to remove its observers when it is marked to be deleted.  The default implemenetation simplies calls deleteLater.