ENH: Include the Attribute Resource when determining Associatable Objects
In the original implementation the attribute resource that the "owns" the
reference item was not being taken into consideration when determining
what could be associated to the item when dealing with the case where the attribute
resource is explicitly associated to other resources.

This modification changes it so that it will include the attribute resource
even if there is no reosurce manager.