Fix entity mapper's input to mirror the represenation's input
In order to get consistent ordering for cell selection (and, therefore,
point picking), we must construct a multiblock dataset that has the
same hierarchical structure as the input data set. We don't want to
have component or instance rendering from our entity mapper, though, so
we simply leave those blocks out of our new dataset.

This fix will likely need to be revisited once we get to prototype
selection.