Instantiate managed operations' parameters prior to observation.
Parameters are constructed lazily, allowing for RAII while having derived
classes construct parameters that are tailored to their use. This can
cause a race condition when observers that are called on a different
thread access parameters at the same time as the thread that created the
operation. Since only managed operations are observed, we avoid this
issue by accessing the parameters as they are created by the manager.