AssignColorsView: fix pqPalletChooser behavior on apply.
Paraview's pqPalletChooser has a weird semi-modal behavior
where it disables interaction with the UI except for the
render view, when it is open. Instead of destroying/recreating
the dialog, call close() so it properly cleans up the
event filter that causes this behavior. Makes the dialog
testable using the recorded xml tests.

add a few more UI names.