ResourceBrowser: make PhraseModel use a factory for construction
Add smtk::view::Manager and Registrar, to allow construction of
PhraseModel and sub-classes by typename. ViewManager is added
to the Wrapper, and retrieved by the ResourcePanel to allow
construction of a qtUIManager with a ViewManager pointer.

Then qtResourceBrowser can use ViewInfo's qtUIManager to retrieve
a ViewManager and dynamically construct a ResourcePhraseModel by
its typename, retrieved from the .json config.