SubphraseGenerator factory: construct based on typename
Add SubphraseGenerators to the view::Manager and Registrar, so
they can be constructed by typename. The default .json config uses
"default" for the type, so that the app settings can choose between
tree and two-level, like the current functionality. This means
the type must be saved by the pqSMTKResourceBrowser to check for
changes.