Remove observation site for operation creation
Operation creation is often performed within observation events (so
the created operation can be added to the launcher queue), so it is
possible to deadlock ModelBuilder by repeatedly create operations when
an operation is running. Since operation creation is not used by any
consuming code written by us or our clients, it is easier to simply
remove the observe call associated with it.