Doc pdfs: Use PNG for actdiag output, avoid crash
When actdiag uses PDF output, it issues a warning, which crashes
when using Sphinx 2, because builder.warn() is no longer available.
The warning is about not finding a font, so an alternative is to
set actdiag_fontpath to a true-type font, but we don't have one
readily available that we are already using. VTK has some in testing
data... PNG seemed easiest, and looks fine.