ENH: Added the concept of an Application String to ItemDefinition
This string is not displayed in the GUI but allows an Application/Workflow
specific string to be attached to the ItemDefinition.  This string is saved in
both the XML and JSON format.

Here is a XML Example:

       <String Name="sA0" Label="sA0" Version="0" OkToInheritCategories="true" CategoryCheckMode="Any" ApplicationString="Application Data Test" NumberOfRequiredValues="1" />

Note that in the future this functionality will be replaced by adding Property support to Items, Definitions, and ItemDefinitions