Fix a polygon selection/test bug.
Polygon models (collections of faces, edges, verts) were being
provided with an empty VTK data object in the new geometry backend.
This caused selection/highlighting of the model to appear to have
no effect since the blank polydata was rendered in the selection
or hover color, rather than the model's cell collection.