Add API to query links regardless of validity
Links connect two persistent objects together via their UUID. This
connection persists even when the RHS is not in memory. A consequence of
this design choice is the possibility for "stale" links to persist, since
there is no automatic link cleanup. This MR adds methods to query a link's
RHS information, regardless of the validity of the link. This allows code
that removes objects to clean up links even after the object has been
removed.