ENH: Adding Regex support for attribute names
Using the AttributeNameRegex option you can now force the attribute name to conform to a regular expression.

NOTE: Currently the expression string should end with an '*'.  If it does not the user will be able to initially enter invalid characters the first time but the widget will not allow the change to be saved.  Subsequent edits will work as expected.  This seems to be a bug in Qt.
Example SBT Change:

    <View Type="Attribute" Title="A" Label="A Atts"
      SearchBoxText="Search by name..." AttributeNameRegex="[a-zA-Z_][a-zA-Z0-9\-]*">
      <AttributeTypes>
        <Att Type="A" />
      </AttributeTypes>
    </View>