BUG: AttributeView improperly reporting validity
The issue was due to the conversion from qtTableWidget to qtTableView. when asking for the item in attribute table view for the status column, the assumption was that it should be null if there was no alter icon and this could be used to indicate that the attribute was valid.  Ths was not the case. The new logic now checks to see if there is a QIcon in the status column to determine if status of an Attribute.