BUG: qtAttributeView Selection Issue
The issue was when either reading in from an existing attribute resource or when switching between attribute resources, the View would show an attribute selected but would not show its contents.  The only solution was to select another attribute and then select the first again.

The problem was introduced when QTableWidget was replaced by qTableView.  The logic still attempted to select the 0th row of the table which selected teh table's item (0,0) which pointed to the status of the attribute instead of item(0,1) which refered to the attribute itself.