Add "Rejects" filter for ReferenceItemDefinition
ReferenceItemDefinition now has a "Rejects" list. While a nonempty "Accepts"
list requires an object to be listed (either directly or indirectly via
inheritance) in order to be valid, any object that is listed (again, directly
or indirectly) in the "Rejects" list is invalid. If an item is listed in both
the "Accepts" and "Rejects" filters (directly or indirectly), the "Rejects"
behavior takes precedence and the object is invalid.