ENH:: Adding the concept of ForceRequired
An item, whose Definition indicates that it is optional, may be forced to be required
based on external factors in a simulation workflow.  This merge request adds the following to Item:

* setForceRequired()
* forceRequired()

Updated IO, JSON, Pybind, and Qt code to support the change.

IO, JSON, Pybind, and Qt Widgets have been updated to support this new functionality.