BUG: Tiled Group Views not properly reporting status
isValid was only looking at children views in Tabs.  Also need to properly connect a childView's
modified signal to the view's childModified slot for tiled and box styles.

There was also an issue with qtAttribteView returning an incorrect not valid when the underlying
association widget is hidden (cause there are no attributes in the list) but its marked as invalid
cause that was status based on the last attribute deleted.