Tell Qt logger to only print message
Qt 5.15 changed the output format to include the category and file/line
information. SMTK already handles this information, so the duplicate (and
often erroneous) information printed to the output was confusing. This MR
reverts Qt's output format to printing only the message, so SMTK can handle
the contents of the message.