Fix crash when saving resource after pipeline Delete
If the user uses the paraview pipeline "Delete" button to remove
an SMTK resource, they are given the chance to save a modified resource.
This then crashes, because the ActiveSource is null, even though we tried
to set it.

Instead, explicitly pass the resource, and then retrieve the active
operation manager to use the WriteResource operator.