Expression parser and evaluator
Evaluators are an addition to SMTK's Attribute system that are a
mechanism to process an Attribute that is an expression. ValueItem
already supported linkage of Attributes to a ValueItem by means of the
<ExpressionType> SBT tag, however Evaluators provide a means to
transform the Items in those expression Attributes into meaningful
results inside SMTK, as opposed to consumers of an Attribute Resource.
These results can be shown to the user in Qt components and serialized
by consuming application code.