python: add SMTK's DLL directory to the Python DLL search path
With Python 3.8, CPython on Windows started restricting DLL searches to
specified directories rather than relying on the PATH environment
variable. This commit adds SMTK's DLL directory to the search path, but
does not add the PATH entries required for dependent libraries. This
works when all of the dependencies share a prefix (as in packages or a
unified install tree), but does not work in a build tree. Considerations
for how to add the required directories for a build tree need to be
done.

CPython change in question:

   https://github.com/python/cpython/commit/2438cdf0e932a341c7613bf4323d06b91ae9f1f1