Race condition on updating the ui
Moved placement of when an attribute was being aded or removed
until after them item was added or removed from the table.
Otherwise the view was being updated causing unexpected behavior
of duplication or what appeared to be an extra removal.  Changing
views causing a redraw would fix the duplication or extra removal.