Registry: hold a strong reference to managers
Now that all registrations occur through this class, a strong reference
should be kept to ensure that if a registration has happened, it is
unregistered before the manager can be destructed. This is necessary
because the `void*` tracking can become confused if a manager is
allocated to the same memory location as an older one that was
deallocated.